+++ src/jsoncons_ext/jsonpath/jsonpath_filter.hpp
@@ -66,120 +66,120 @@ public:
     }
     virtual bool accept_single_node() const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json evaluate_single_node() const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json exclaim() const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool eq_term(const term&) const 
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool eq(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json ne_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool ne(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json regex_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool regex2(const string_type&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool ampamp_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool ampamp(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool pipepipe_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool pipepipe(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool lt_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool lt(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool gt_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual bool gt(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 
     virtual Json minus_term(const term&) const 
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json  minus(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 
     virtual Json left_minus(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 
     virtual Json unary_minus() const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json plus_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json  plus(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json mult_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json mult(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 
     virtual Json div_term(const term&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
     virtual Json div(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 
     virtual Json left_div(const Json&) const
     {
-        throw parse_exception(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
+        throw parse_error(jsonpath_parser_errc::invalid_filter_unsupported_operator,1,1);
     }
 };
 
@@ -945,9 +945,9 @@ public:
             return t.operand().evaluate_single_node();
 
         }
-        catch (const parse_exception& e)
+        catch (const parse_error& e)
         {
-            throw parse_exception(e.code(),line_,column_);
+            throw parse_error(e.code(),line_,column_);
         }
     }
 
@@ -958,9 +958,9 @@ public:
             auto t = evaluate(context_node,tokens_);
             return t.operand().accept_single_node();
         }
-        catch (const parse_exception& e)
+        catch (const parse_error& e)
         {
-            throw parse_exception(e.code(),line_,column_);
+            throw parse_error(e.code(),line_,column_);
         }
     }
 };
@@ -1138,7 +1138,7 @@ public:
                         }
                         else
                         {
-                            throw parse_exception(jsonpath_parser_errc::unexpected_operator,line_,column_);
+                            throw parse_error(jsonpath_parser_errc::unexpected_operator,line_,column_);
                         }
                         break;
                     }
@@ -1241,7 +1241,7 @@ public:
                 case ' ':case '\t':
                     break;
                 default:
-                    throw parse_exception(jsonpath_parser_errc::invalid_filter,line_,column_);
+                    throw parse_error(jsonpath_parser_errc::invalid_filter,line_,column_);
                     break;
                 }
                 ++p;
@@ -1275,9 +1275,9 @@ public:
                                     auto val = Json::parse(buffer);
                                     add_token(token<Json>(token_type::operand,std::make_shared<value_term<Json>>(val)));
                                 }
-                                catch (const parse_exception& e)
+                                catch (const parse_error& e)
                                 {
-                                    throw parse_exception(e.code(),line_,column_);
+                                    throw parse_error(e.code(),line_,column_);
                                 }
                                 buffer.clear();
                             }
@@ -1292,9 +1292,9 @@ public:
                                 auto val = Json::parse(buffer);
                                 add_token(token<Json>(token_type::operand,std::make_shared<value_term<Json>>(val)));
                             }
-                            catch (const parse_exception& e)
+                            catch (const parse_error& e)
                             {
-                                throw parse_exception(e.code(),line_,column_);
+                                throw parse_error(e.code(),line_,column_);
                             }
                             buffer.clear();
                         }
@@ -1319,9 +1319,9 @@ public:
                                 auto val = Json::parse(buffer);
                                 add_token(token<Json>(token_type::operand,std::make_shared<value_term<Json>>(val)));
                             }
-                            catch (const parse_exception& e)
+                            catch (const parse_error& e)
                             {
-                                throw parse_exception(e.code(),line_,column_);
+                                throw parse_error(e.code(),line_,column_);
                             }
                             buffer.clear();
                         }
@@ -1364,9 +1364,9 @@ public:
                                 auto val = Json::parse(buffer);
                                 add_token(token<Json>(token_type::operand,std::make_shared<value_term<Json>>(val)));
                             }
-                            catch (const parse_exception& e)
+                            catch (const parse_error& e)
                             {
-                                throw parse_exception(e.code(),line_,column_);
+                                throw parse_error(e.code(),line_,column_);
                             }
                             buffer.clear();
                         }
@@ -1409,9 +1409,9 @@ public:
                                 auto val = Json::parse(buffer);
                                 add_token(token<Json>(token_type::operand,std::make_shared<value_term<Json>>(val)));
                             }
-                            catch (const parse_exception& e)
+                            catch (const parse_error& e)
                             {
-                                throw parse_exception(e.code(),line_,column_);
+                                throw parse_error(e.code(),line_,column_);
                             }
                             buffer.clear();
                         }
@@ -1532,7 +1532,7 @@ public:
                     }
                     break;
                 default: 
-                    throw parse_exception(jsonpath_parser_errc::invalid_filter,line_,column_);
+                    throw parse_error(jsonpath_parser_errc::invalid_filter,line_,column_);
                     break;
                 };
                 break;
@@ -1560,7 +1560,7 @@ public:
                     }
                     break;
                 default: 
-                    throw parse_exception(jsonpath_parser_errc::invalid_filter,line_,column_);
+                    throw parse_error(jsonpath_parser_errc::invalid_filter,line_,column_);
                     break;
                 };
                 ++p;
@@ -1636,7 +1636,7 @@ public:
                 case ' ':case '\t':
                     break;
                 default: 
-                    throw parse_exception(jsonpath_parser_errc::invalid_filter_expected_slash,line_,column_);
+                    throw parse_error(jsonpath_parser_errc::invalid_filter_expected_slash,line_,column_);
                     break;
                 };
                 ++p;
@@ -1684,7 +1684,7 @@ public:
         }
         if (depth != 0)
         {
-            throw parse_exception(jsonpath_parser_errc::invalid_filter_unbalanced_paren,line_,column_);
+            throw parse_error(jsonpath_parser_errc::invalid_filter_unbalanced_paren,line_,column_);
         }
         *end_ptr = p;