+++ doc/md/Changelog.md
@@ -242,7 +242,7 @@ Non-beaking API Changes
 - json member functions `begin_members` and `end_members` have been deprecated, instead use `members().begin()` and `members.end()`
 - json member function `has_member` has been deprecated, instead use `count`. Rationale: consistency with C++ containers
 - json member function `remove_member` has been deprecated, instead use `remove`. Rationale: only member function left with _element or _member suffix 
-- json_parse_exception renamed to parse_exception, json_parse_exception typedef to parse_exception
+- json_parse_exception renamed to parse_error, json_parse_exception typedef to parse_error
 - json::parse(std::istream& is) renamed to json::parse_stream. json::parse(std::istream is) is deprecated but still works.
 
 0.98.2 Release